src/agents/sr/requirements.txt
pandas
streamlit
matplotlib
langchain
langchain-openai
python-dotenv